requirements.txt
streamlit
openai>=1.0.0
langchain==0.1.13
langchain-core==0.1.32
langchain-openai==0.0.8
langchain-community==0.0.30
faiss-cpu
PyPDF2
python-dotenv
pandas
